Add files via upload
My calculator project showcases a sleek and user-friendly interface, built primarily with HTML and CSS, to deliver a seamless arithmetic experience. The HTML structure is meticulously crafted to ensure logical placement of buttons and display areas, providing a solid foundation for user interactions. Complementing the HTML, CSS is employed to enhance the visual appeal and responsiveness of the calculator.